refactor: Resolve circular dependency by creating config module
